refactor: decompose 1175-line pose6dof.py into domain modules (fixes #2386)
