+++ src/shared/python/spatial_algebra/pose6dof/__init__.py
@@ -1,0 +1,32 @@
+from .placement import EntityPlacement, PlacementGroup
+from .pose import Pose6DOF
+from .rotations import (
+    axis_angle_to_rotation_matrix,
+    euler_to_quaternion,
+    euler_to_rotation_matrix,
+    quaternion_inverse,
+    quaternion_multiply,
+    quaternion_to_euler,
+    quaternion_to_rotation_matrix,
+    rotation_matrix_to_euler,
+    rotation_matrix_to_quaternion,
+    slerp,
+)
+from .transform import Transform6DOF
+
+__all__ = [
+    "euler_to_rotation_matrix",
+    "rotation_matrix_to_euler",
+    "euler_to_quaternion",
+    "quaternion_to_euler",
+    "quaternion_to_rotation_matrix",
+    "rotation_matrix_to_quaternion",
+    "axis_angle_to_rotation_matrix",
+    "quaternion_multiply",
+    "quaternion_inverse",
+    "slerp",
+    "Pose6DOF",
+    "Transform6DOF",
+    "EntityPlacement",
+    "PlacementGroup",
+]
